pyproject.toml
[tool.poetry]
name = "house-notes-be"
version = "0.1.0"
description = "Backend that saves notes"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "house_notes_be"}]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.2.2"
flask-sqlalchemy = "^3.0.2"
python-dotenv = "^0.21.0"
flask-migrate = "^4.0.2"
psycopg2 = "^2.9.5"
jsonschema = "^4.17.3"
flask-marshmallow = "^0.14.0"
gunicorn = "^20.1.0"
marshmallow-sqlalchemy = "^0.28.1"
pytest-html = "^3.2.0"
flask-jwt-extended = "^4.4.4"
passlib = "^1.7.4"

[tool.poetry.group.dev.dependencies]
pylint = "^2.15.10"
black = "^22.12.0"
pytest = "^7.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
